third_party/blink/web_tests/external/wpt/html/webappapis/scripting/events/event-handler-attributes-windowless-body-expected.txt

This is a testharness.js-based test.
Found 10 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Return null when getting the pagereveal event handler of a windowless body
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Ignore setting of pagereveal window event handlers on windowless body
  assert_equals: expected (object) null but got (function) function "function() { return "Handler attached to windowless element"; }"
[FAIL] Return null when getting the pageswap event handler of a windowless body
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Ignore setting of pageswap window event handlers on windowless body
  assert_equals: expected (object) null but got (function) function "function() { return "Handler attached to windowless element"; }"
[FAIL] Ignore setting of unload window event handlers on windowless body
  assert_equals: expected (function) function "function () { return "Handler attached to the window"; }" but got (object) null
[FAIL] Return null when getting the pagereveal event handler of a windowless frameset
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Ignore setting of pagereveal window event handlers on windowless frameset
  assert_equals: expected (object) null but got (function) function "function() { return "Handler attached to windowless element"; }"
[FAIL] Return null when getting the pageswap event handler of a windowless frameset
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Ignore setting of pageswap window event handlers on windowless frameset
  assert_equals: expected (object) null but got (function) function "function() { return "Handler attached to windowless element"; }"
[FAIL] Ignore setting of unload window event handlers on windowless frameset
  assert_equals: expected (function) function "function () { return "Handler attached to the window"; }" but got (object) null
Harness: the test ran to completion.
